Add mysql dumps to backups as well as full backups

+++ app/routes/api/backup/backup.js
@@ -2,16 +2,18 @@ const express            = require('express');
 const router             = express.Router();
 const eqemuConfigService = require('../../../core/eqemu-config-service')
 const pathManager        = require('../../../core/path-manager')
+const mysqldump          = require('../../../core/mysqldump-service')
 const path               = require('path')
 const fs                 = require('fs')
 const yazl               = require("yazl");
 const recursive          = require("recursive-readdir");
 const slugify            = require('slugify')
 const util               = require('util')
+const os                 = require('os')
 
 router.get('/:download', async function (req, res, next) {
-  const timeoutThirtyMinutes = 1000 * 60 * 30;
-  res.connection.setTimeout(timeoutThirtyMinutes);
+  const timeoutSixtyMinutes = 1000 * 60 * 60;
+  res.connection.setTimeout(timeoutSixtyMinutes);
 
   const download       = req.params.download;
   const serverLongName = eqemuConfigService.getServerLongName()
@@ -22,7 +24,7 @@ router.get('/:download', async function (req, res, next) {
     {remove: /[*+~()'"!:@]/g}
   )
 
-  let requestedDownloadPath;
+  let requestedDownloadPath = "";
   switch (download) {
     case "quests":
       requestedDownloadPath = download;
@@ -33,25 +35,25 @@ router.get('/:download', async function (req, res, next) {
     case "lua_modules":
       requestedDownloadPath = download;
       break;
+    case "database":
+      break;
+    case "full":
+      requestedDownloadPath = "./";
+      break;
     default:
       res.json({error: 'Invalid download requested'})
   }
 
-  const realPath = path.join(pathManager.getEmuServerPath(), requestedDownloadPath)
-  const zipFile  = path.join(pathManager.getEmuServerPath(), archiveName)
-  let zip        = new yazl.ZipFile();
-
-  recursive(realPath, function (err, files) {
-    files.forEach(function (file) {
-      var zipPath = file.replace(realPath + "/", "")
-      zip.addFile(file, zipPath);
-      console.debug("[backup.js] Adding file [%s] to zipPath [%s]", file, zipPath)
-    });
+  if (download === "database") {
+    const dumpFile = mysqldump.dump()
+    const zip      = new yazl.ZipFile();
+    const zipFile  = path.join(os.tmpdir(), path.basename(dumpFile) + ".zip")
+    zip.addFile(dumpFile, path.basename(dumpFile));
 
     zip.outputStream.pipe(
       fs.createWriteStream(zipFile)
     ).on("close", function () {
-      console.debug("zip file downloaded: " + zipFile);
+      fs.unlinkSync(dumpFile);
 
       res.download(zipFile, path.basename(zipFile), function (err) {
         fs.unlinkSync(zipFile);
@@ -59,7 +61,66 @@ router.get('/:download', async function (req, res, next) {
     });
 
     zip.end();
-  });
+  }
+
+  if (requestedDownloadPath !== "") {
+    const realPath = path.join(pathManager.getEmuServerPath(), requestedDownloadPath)
+    const zipFile  = path.join(os.tmpdir(), archiveName)
+    let zip        = new yazl.ZipFile();
+
+    recursive(realPath, function (err, files) {
+      files.forEach(function (file) {
+        const baseServerFilePath          = file.replace(pathManager.getEmuServerPath(), "")
+        const firstLevelFile              = baseServerFilePath.split("/")[0]
+        const ignoreBackupFirstLevelFiles = [
+          'backups',
+          'eqemu-web',
+          'eqemu-admin',
+          'db_update',
+          'updates_staged',
+          'logs',
+          'shared',
+          'export'
+        ];
+
+        if (ignoreBackupFirstLevelFiles.includes(firstLevelFile)) {
+          console.debug("[backup.js] Skipping file/folder [%s]", firstLevelFile)
+          return;
+        }
+
+        const baseFile = path.basename(file)
+        if (baseFile.includes(".zip")) {
+          console.debug("[backup.js] Skipping .zip file [%s]", baseFile)
+          return;
+        }
+
+        let zipPath = file.replace(pathManager.getEmuServerPath(), "")
+        console.debug("[backup.js] Adding file [%s] to zipPath [%s]", file, zipPath)
+        zip.addFile(file, zipPath);
+      });
+
+      if (download === "full") {
+        const dumpFile = mysqldump.dump()
+        zip.addFile(dumpFile, path.join("database", path.basename(dumpFile)));
+      }
+
+      zip.outputStream.pipe(
+        fs.createWriteStream(zipFile)
+      ).on("close", function () {
+        console.debug("zip file downloaded: " + zipFile);
+
+        res.download(zipFile, path.basename(zipFile), function (err) {
+          fs.unlinkSync(zipFile);
+          if (download === "full") {
+            fs.unlinkSync(mysqldump.getBackupFullSqlFilePath());
+          }
+        });
+      });
+
+      zip.end();
+    });
+  }
+
 });
 
 module.exports = router;
